refactor(keyboard): move and rename TrimeKeyEffects
